machine_learning/project/text-service/pyproject.toml
[project]
name = "text-service"
version = "0.1.0"
description = "Micrservice for process messages"
authors = [
    {name = "matebookWindows",email = "dan.kovalev2013@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "dishka (>=1.5.3,<2.0.0)",
    "faststream[kafka] (>=0.5.39,<0.6.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "openai (>=1.75.0,<2.0.0)",
    "deep-translator (>=1.11.4,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "concurrent-log-handler (>=0.9.25,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
